Move MrpRotationConfig from Types.h to Algorithm.h
